requirements.txt
altair==5.1.1
asn1crypto==1.5.1
attrs==23.1.0
Babel==2.12.1
bandit==1.7.7
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
dacite==1.8.1
dbt-metadata-client==0.1.0
duckdb==0.9.0
Faker==19.6.2
favicon==0.7.0
filelock==3.12.4
flake8==6.1.0
fonttools==4.43.0
gitdb==4.0.10
GitPython==3.1.37
graphql-core==3.2.3
htbuilder==0.6.2
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
multimethod==1.10
mypy-extensions==1.0.0
networkx==3.1
numerize==0.12
numpy==1.23.5
oscrypto==1.3.0
packaging==23.1
pandas==2.0.3
pandas-profiling==3.6.6
pathspec==0.11.2
patsy==0.5.3
pbr==6.0.0
phik==0.12.3
Pillow==10.0.1
platformdirs==3.8.1
plotly==5.17.0
protobuf==4.24.3
pyarrow==10.0.1
pycodestyle==2.11.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==1.10.13
pydeck==0.8.1b0
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.3
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
scipy==1.11.3
seaborn==0.12.2
sgqlc==16.3
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.2.0
sortedcontainers==2.4.0
soupsieve==2.5
st-annotated-text==4.0.1
statsmodels==0.14.0
stevedore==5.1.0
streamlit==1.27.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.2
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-pandas-profiling==0.1.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
validators==0.22.0
visions==0.7.5
wordcloud==1.9.2
ydata-profiling==4.5.1
zipp==3.17.0
